+++ app/datatables/elements_datatable.rb
@@ -13,9 +13,9 @@ class ElementsDatatable < IndexDatatable
       competition_class: "competitions.competition_class",
       competition_type: "competitions.competition_type",
       season: "competitions.season",
-      category: "scores.category",
-      segment: "scores.segment",
-      segment_type: "scores.segment_type",
+      category: "categories.name",
+      segment: "segments.name",
+      segment_type: "segments.segment_type",
       #date: "scores.performed_starting_time.starting_time",
       date: "scores.date",
       skater_name: "skaters.name",
@@ -37,11 +37,11 @@ class ElementsDatatable < IndexDatatable
       columns[:name].operator = params[:name_operator].presence || :matches
       columns[:goe].operator = params[:goe_operator].presence || :eq
     end
-    
+    columns[:category].operator = :eq    
     default_orders([[:value, :desc]])
   end
 
   def fetch_records
-    Element.includes(:score, score: [:competition, :skater]).references(:score, score: [:competition, :skater]).all
+    Element.includes(:score, score: [:competition, :skater, :category, :segment]).references(:score, score: [:competition, :skater, :category, :segment]).all
   end
 end