+++ db/migrate/002_create_competitions.rb
@@ -22,7 +22,8 @@ class CreateCompetitions < ActiveRecord::Migration[5.1]
     ################
     # category results
     create_table :category_results do |t|
-      t.string :category
+      #t.string :category
+      t.belongs_to :category
       t.integer :ranking
       t.float :points
 
@@ -39,8 +40,10 @@ class CreateCompetitions < ActiveRecord::Migration[5.1]
     ################
     # performed segments
     create_table :performed_segments do |t|
-      t.string :category
-      t.string :segment
+      #t.string :category
+      t.belongs_to :category
+      #t.string :segment
+      t.belongs_to :segment
       t.datetime :starting_time, default: Time.new(1970, 1, 1, 0, 0, 0)
       
       ## relations