+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - help for rake:update
 
 - skater updater spec: detail
+- uniq list: include
 
 ### new feature
 
@@ -24,6 +25,7 @@
 
 ## 1.0.6
 - element subtype
+- category, segment as active record
 
 ## 1.0.5
 - category result has short and free